Change Player controller for online
Changed player controller class to support multiplayer.

+++ gd4_sfml_tcp_game/MultiplayerState.hpp
@@ -16,7 +16,7 @@ public:
 
 private:
 	void UpdateBroadcastMessage(sf::Time elapsed_time);
-	void HandlePacket(sf::Int32 packet_type, sf::Packet& packet);
+	void HandlePacket(sf::Int16 packet_type, sf::Packet& packet);
 
 private:
 	sf::RenderWindow& m_window;
